Update dependencies, enhance API functionality, and improve event handling
- Updated `tonic`, `prost`, and related dependencies in `Cargo.toml` and `Cargo.lock` for better compatibility.
- Refactored `build.rs` to utilize `tonic_prost_build` for gRPC service generation.
- Enhanced WebSocket event broadcasting for message delivery and configuration updates.
- Improved API routes for handling messages, notes, and delegations with better error handling and response structures.
- Added new methods in `DbClient` for artifact and delegation operations, ensuring better observability and functionality.

These changes improve the overall API structure, maintainability, and responsiveness to events while ensuring compatibility with updated dependencies.

+++ caliber-tui/src/widgets/progress.rs
@@ -33,7 +33,7 @@ impl ProgressBar {
         let gauge = Gauge::default()
             .block(Block::default().title(self.title.as_str()).borders(Borders::ALL))
             .gauge_style(style)
-            .ratio(ratio);
+            .ratio(ratio as f64);
         f.render_widget(gauge, area);
     }
 }
